refactor: rename Lookup namespace to Map
