+++ src/Yaapii.Atoms/Func/BowAction.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Yaapii.Atoms.Lookup;
+using Yaapii.Atoms.Map;
 
 namespace Yaapii.Atoms.Func
 {
