+++ src/Yaapii.Atoms/Map/FallbackMap.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using Yaapii.Atoms.Fail;
 using Yaapii.Atoms.Scalar;
 
-namespace Yaapii.Atoms.Lookup
+namespace Yaapii.Atoms.Map
 {
     /// <summary>
     /// Map which can return a fallback value generated by a fallback function
