+++ src/Yaapii.Atoms/Map/Joined.cs
@@ -23,7 +23,7 @@
 using System.Collections.Generic;
 using Yaapii.Atoms.Enumerable;
 
-namespace Yaapii.Atoms.Lookup
+namespace Yaapii.Atoms.Map
 {
     /// <summary>
     /// Joined map.
