+++ tests/Yaapii.Atoms.Tests/Map/MapEnvelopeTest.cs
@@ -25,7 +25,7 @@ using System;
 using System.Collections.Generic;
 using Xunit;
 
-namespace Yaapii.Atoms.Lookup.Tests
+namespace Yaapii.Atoms.Map.Tests
 {
     public class MapEnvelopeTest
     {
